refactor(ladderSize): LadderSize 를 제거하고 Ladder에서 크기를 관리하도록 변경
